バックエンドに送るペイロードを整形
