+++ react/src/types/pages.ts
@@ -16,3 +16,14 @@ export type MainViewProps = {
   readonly onRequestCreate: () => void;
   readonly onRequestUpdate: () => void;
 };
+
+export type Row = {
+  date: string; // "YYYY-MM-DD"
+  start: string; // "HH:MM"
+  end: string; // "HH:MM"
+};
+
+export type TimeslotPayload = {
+  start_time: string; // "YYYY-MM-DD HH:MM:SS"
+  end_time: string; // "YYYY-MM-DD HH:MM:SS"
+};
